Refactoring themes #35
* split files for themes
