--- src/context/ThemeContext/theme.ts
@@ -1,96 +1,0 @@
-import { createTheme } from '@mui/material/styles';
-import { Components } from '@mui/material/styles/components';
-import { Theme } from '@mui/material/styles/createTheme';
-
-const components: Components<Omit<Theme, 'components'>> = {
-  MuiButton: {
-    defaultProps: {
-      variant: 'contained',
-      fullWidth: true,
-    },
-  },
-};
-
-export const darkTheme = createTheme({
-  components: {
-    ...components,
-    // Scrollbar
-    MuiCssBaseline: {
-      styleOverrides: {
-        body: {
-          scrollbarColor: '#6b6b6b #2b2b2b',
-          '&::-webkit-scrollbar, & *::-webkit-scrollbar': {
-            backgroundColor: '#2b2b2b',
-          },
-          '&::-webkit-scrollbar-thumb, & *::-webkit-scrollbar-thumb': {
-            borderRadius: 8,
-            backgroundColor: '#6b6b6b',
-            minHeight: 24,
-            border: '3px solid #2b2b2b',
-          },
-          '&::-webkit-scrollbar-thumb:focus, & *::-webkit-scrollbar-thumb:focus': {
-            backgroundColor: '#959595',
-          },
-          '&::-webkit-scrollbar-thumb:active, & *::-webkit-scrollbar-thumb:active': {
-            backgroundColor: '#959595',
-          },
-          '&::-webkit-scrollbar-thumb:hover, & *::-webkit-scrollbar-thumb:hover': {
-            backgroundColor: '#959595',
-          },
-          '&::-webkit-scrollbar-corner, & *::-webkit-scrollbar-corner': {
-            backgroundColor: '#2b2b2b',
-          },
-        },
-      },
-    },
-  },
-  palette: {
-    mode: 'dark',
-  },
-});
-
-export const lightTheme = createTheme({
-  components: {
-    ...components,
-    // Scrollbar
-    MuiCssBaseline: {
-      styleOverrides: {
-        body: {
-          scrollbarColor: '#c1c1c1 #f1f1f1',
-          '&::-webkit-scrollbar, & *::-webkit-scrollbar': {
-            backgroundColor: '#f1f1f1',
-          },
-          '&::-webkit-scrollbar-thumb, & *::-webkit-scrollbar-thumb': {
-            borderRadius: 8,
-            backgroundColor: '#c1c1c1',
-            minHeight: 24,
-            border: '3px solid #f1f1f1',
-          },
-          '&::-webkit-scrollbar-thumb:focus, & *::-webkit-scrollbar-thumb:focus': {
-            backgroundColor: '#a8a8a8',
-          },
-          '&::-webkit-scrollbar-thumb:active, & *::-webkit-scrollbar-thumb:active': {
-            backgroundColor: '#a8a8a8',
-          },
-          '&::-webkit-scrollbar-thumb:hover, & *::-webkit-scrollbar-thumb:hover': {
-            backgroundColor: '#a8a8a8',
-          },
-          '&::-webkit-scrollbar-corner, & *::-webkit-scrollbar-corner': {
-            backgroundColor: '#f1f1f1',
-          },
-        },
-      },
-    },
-  },
-  palette: {
-    mode: 'light',
-  },
-});
-
-declare module '@mui/material/styles' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface Theme {}
-
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface ThemeOptions {}
-}
